rewrite readdir prefetching logic
Summary:
Bring back readdir prefetching, which was disabled as part of the
edenapi aux metadata rollout, and tighten up its logic to accelerate
specific use cases.

To optimize `find .`, prefetch subtrees when a subtree is stat().

To optimize `find . -empty`, prefetch blob metadata when a file is
stat().

Add a config knob to opt out of blob prefetching, tree prefetching, or
both.

Reviewed By: xavierd

Differential Revision: D41701675

fbshipit-source-id: 29b9c28f0f09cf686e3f256bd8c993e9b466e012