commit_graph: query methods should fail if any commit doesn't exist
Summary: The commit graph `is_ancestor` and `get_ancestors_difference` methods both omit any parameters that don't exist.  In order to implement fallback, we need these to fail if any parameter doesn't exist.  This will also be better for correctness in the long run: once the table is fully populated there should be no reason any client asks about a commit that doesn't exist, so it's better to fail if they do.

Reviewed By: YousefSalama

Differential Revision: D42353987

fbshipit-source-id: 247d83b91c23940bc900b3ed4561f48c46fd49f5