ODS counters for surfacing config refresh failures
Summary: During hot-reloading of configuration, any failures encountered are simply logged and the process continues since a config refresh failure should not halt the actual service. However, this could also result in the failures getting silently ignored. This diff adds ODS metrics that can be used to configure alerts that would be fired in case of failures.

Reviewed By: mitrandir77

Differential Revision: D41492081

fbshipit-source-id: dc55b9a95c3dee505b4dc86c8f6717774b6ebac4