config: respect config filters for static config
Summary:
The config "Options" object includes filter functions that can arbitrarily mutate or filter out any config value as the config is being loaded. One important place filters are used is to drop command alias configuration when running under HGPLAIN mode. (Note that I feel it would be better to not mutate the config and instead skip the aliases during command dispatch, but I didn't attempt to make that change.)

The recent static config improvements accidentally skirt the config filters. This meant aliases were respected with HGPLAIN=1. For example, this caused "jf submit -s" to pop open an editor because jf runs "HGPLAIN=1 hg metaedit ..." and there is a built in alias for metaedit => "metaedit --batch", and "--batch" forces an editor to pop up when there is more than one revision to edit.

Fix by computing static config "overrides" based on anything that is filtered out. Any overrides are layered in on top of the static configs.

Reviewed By: zzl0

Differential Revision: D42395024

fbshipit-source-id: 2f5b4420f3583bab9cc642aae8e3c1f1d0ff952e