(build): Support sapling in tag-name.sh
Summary: Currently, tag-name.sh only works with git.  This is fine, because
it's only used in GitHub CI, but I also want to use this script for local
builds (which will be in a future commit).  This commit will use either sl or
git for creating tag names.

Test-Plan: Automated testing blocked on #447

```sh
❯ ./ci/tag-name.sh
0.2.20230107-190233-hc1336fcb
```