inodes: when invalidating only stat when cutoff is valid
Summary:
Manually testing GC, I noticed that when I try to invalidate everything after
building EdenFS, I'm left with 10k loaded files, even though I would expect
that number to be close to 0. What I believe is happening is that in some
cases, the stat call will lay placeholder on disk, which will force the inode
to be loaded (inside PrjfsDispatcherImpl::lookup) and its refcount incremented
in the background. This happens concurrently with the invalidation, and if the
load happens after the invalidation, files/trees will be re-loaded and their
refcount incremented, leading to what I'm observing.

Some users also have reported invalidation taking a really long time, seemingly
due to all the stats calls, this should aleviate this too.

Reviewed By: kmancini

Differential Revision: D42467112

fbshipit-source-id: e1f242dc3f53b9dc34e1a24f99f13f78636a4b52