Platform API for copying to clipboard
Summary:
We saw reports of copying diff numbers to the clipboard not working in the android studio embedding of ISL. This seems to be a permissions issue with android studio webviews not allowing `navigator.clipboard.writeText`.

Two options to fix this:
- see if there's some API for enabling this permission
- add a JS bridge API so that you can do the clipboard copy from java/kotlin instead of in javascript.

This diff adds an ISL UI platform api for `clipboardCopy`. This can be used by android studio to implement the copy. I put in a placeholder implementation as well.

Reviewed By: bolinfest

Differential Revision: D42440582

fbshipit-source-id: 3de304f6e4cec03eddb700850b251617c54fe186