Add server-side support to delete files from the file system
Summary:
Following up from the last diff, here we add server side support for deleting files, so that you can delete untracked files in the UI.

We actually need to be a little careful with this for security. We don't really want to let the client send a message to the server to delete ANY file on the server machine. We shouldn't assume the client is completely trusted. Instead, we can limit it to only allow deleting files from inside the given repository. This should be a lot safer.

Reviewed By: bolinfest

Differential Revision: D42443915

fbshipit-source-id: 32a8addf4e70c5a19c6ffe3cce40c49976b961c6