Add "addremove" button
Summary:
If you have a lot of untracked changes, we can run `sl addremove` to add them in one go. This also works for missing files, to mark them as removed.

Is "Add/Remove" a good enough name for the button? In old vscode, it's "Add Untracked/Remove Missing", which is more accurate but a super long button

Reviewed By: bolinfest

Differential Revision: D42442739

fbshipit-source-id: 91e1ec387a7ca51616657674c65cfd74add95b2f