Forget button for added files, Delete button for untracked files
Summary:
We had a "revert" button on added files. This was kind of confusing / not intuitive, since it doesn't delete the file, it just stops tracking it.

I think making this explicit with a "forget" button is a lot easier to understand.

Furthermore, untracked files could not be deleted in any way.

My use case, which I think is common, is that I made a file, added it to `sl` with `sl add`, then later realized I don't need that file. So from ISL I'd like to be able to just delete the whole file.

Turns out `sl` doesn't have any command which will delete an added file from disk (understandably, we don't want to delete files that aren't being tracked, it's a lossy operation).

But as a user, it's nice to be able to tell ISL to delete the file if I'm sure I don't need it. I don't want to drop to the CLI just to run `rm` for one file.

So we need another action that deletes the file. In this diff I add the button, but only hook it up in the next diff.

Reviewed By: bolinfest

Differential Revision: D42443914

fbshipit-source-id: 50959f9385d79ab3305269ce920236cc331ef0fb