disable min speed check for commit graph requests
Summary:
disable min speed check for commit graph requests.

for commit graph request, server returns result once it is fully ready, due to phases calculation and derivation filenodes for public roots, so the check must be disabled.

we still have a general timeout for an edenapi request, that is configured as 30 min, so enough for hg pull commands to complete for all edge cases, including aosp repo pulls of other main lines.

Reviewed By: muirdm

Differential Revision: D42478560

fbshipit-source-id: ff73cafee78175e07bc6379658f081c393aa91d3