support abbreviated git commits in SCS
Summary:
Since [the rollout](https://fb.workplace.com/groups/3112141815538647/permalink/5766056630147139/) of SCM query proxy redirection to SCS for git commits, we have seen that queries that use a shorthand for a git rev fail in SCS and use the fallback path.

This is because commit prefix resolution was not implemented in SCS for the GIT identity scheme.

This diff addresses that disparity by:
* adding a type: `GitSha1Prefix` that can return the min and max changeset for a given prefix
* implementing `get_in_range` on bonsai_git_mapping, which queries the `bonsai_git_mapping` database for any matching commit in the range defined by a `GitSha1Prefix`
* implementing the necessary boilerplate to call this implementation when the `repo_resolve_commit_prefix` endpoint on SCS is called with a the GIT identity scheme specified

The implementation follows the same pattern as the prefix resolution for Mercurial.

Reviewed By: yancouto

Differential Revision: D42102242

fbshipit-source-id: fc688ac28ce4c0013f792933eb3a1732e8918768