blobrepo: Remove blobrepo from CommitSyncer
Summary:
The struct `CommitSyncRepos` and other related structs and enums (in and around `eden/mononoke/commit_rewriting/cross_repo_sync/src/lib.rs`) used to have direct references to `BlobRepo`. This diff changes that to use a template instead.

I needed to fix a bunch of binaries that used this.

Differential Revision: D42477257

fbshipit-source-id: f9811e365aeb42fed6f8384a232257b43ef41282