commit_graph: add backfill-one subcommand to backfiller
Summary:
Adds a backfill-one subcommand to the backfiller, that adds a changeset to the commit graph and all its missing ancestors. Useful when trying to backfill one particular changeset.

Context: The backfiller assumes that changesets in the old changesets table are sorted topologically, but when backfilling www it encountered one changeset that had a lower auto-increment id than its parent. This subcommand allowed backfilling the missing parent to unblock the backfiller.

Reviewed By: Croohand

Differential Revision: D42535521

fbshipit-source-id: 7a1bdb41de9b611f9308a80b06ffe96bc8dfbe0c