commit_graph: add get_ancestors_frontier_with method
Summary:
Adds a method that takes a vector of changeset ids and a property closure, and returns a frontier for the set of all ancestors of those changesets that satisfy the property closure.

The property closure needs to be monotonic i.e. if the property holds for one changeset it needs to hold for all its parents.

Reviewed By: mitrandir77

Differential Revision: D42477200

fbshipit-source-id: f2963bcaf983dba92370f31e5106f6e014f27bce