return phases from server if they are known
Summary:
return phases from server if they are known

As we discussed, with edenapi strait after the "commitgraph" request we need to
ask the sever for the mutations for draft commits. Computing drafts on client
side is ok but extra overhead since the phases have been calculated already
server side for derivation purposes (derive filenodes for public roots to
support linknodes). It would be nice to return them, so the client can use if
they want.

Reviewed By: RajivTS

Differential Revision: D42451892

fbshipit-source-id: db94a6b14b9a8e1966d2355806b569d6abee546d