Use ref to CoreContext in Changesets trait
Summary:
I changed both traits `Changesets` and `ChangesetFetcher` to take context by reference instead of by value. This removes a lot of unnecessary cloning we were doing.

It was really annoying because A LOT of places use this. I fixed all of them to get them to compile, but not all of them to do the minimal number of clones, as there's other stuff that still unnecessarily takes context by value.

Reviewed By: mitrandir77

Differential Revision: D42547207

fbshipit-source-id: 2d852abb8bf7825026b3b2ffe647ba0fc9e35b56