glob: use `SerialExecutor` to avoid fan-out
Summary:
To address the Thrift QueueTimeout issue we have been seeing in EdenFS Thrift server.

The underlying issue is that, `globFiles` Thrift method in EdenFS is causing too much fanouts, overloading the Thrift CPU worker. As a result, all the new incoming requests are waiting in the Thrift queue too long to get picked, resulting as QueueTimeout exceptions on the client side.

This diff adds `folly::SerialExecutor` for the `globFiles` method to limit concurrency when processing glob requests.

Reviewed By: chadaustin

Differential Revision: D41390586

fbshipit-source-id: 54d834bb370d4ae693089051f200b7103de2f85b