Start new server if sl version mismatches
Summary:
Right now, our server re-use is a bit aggressive. A spawned ISL server will stick around until you close all ISL windows or kill the server yourself or use --force.

But if we updated sl in the background, then there's room for incompatibilities. The server you're already running would be using the old javascript, but it would server the newer javascript for the client, which could easily cause bugs.

In such a situation, we should pretend the user passed `--force`.

The implementation of this is slightly annoying. We're not currently set up to check for the previous server information before we try spawning the server. So we won't be able to add `--force` directly. Instead, if we fail to spawn, then see the existing instance uses a different version, we'll have to kill it and then restart the entire process. This seems to work well in practice.

Even though we're recursively calling `main`, I don't see a way for it to loop forever, though it is worth making sure that's the case. Even if you had a pathological case of multiple ISL servers trying to spawn at the same time.

TODO: we need to update the isl wrapper inside the python sl launcher to pass the version string in

Reviewed By: bolinfest

Differential Revision: D42561951

fbshipit-source-id: bdd0a6e0fb9065ee5fb7cb3110a5a937fc40d9ce