Log version + platform name on startup
Summary:
Now that we have a notion of version, we should log it during startup. This makes our logs easier to read through.

I think including the platform name would also be worth doing.

A little bit unfortunately, we have the platform name in the client platform, but we really want to log it on the server. So here I also just duplicate the platform name into the server platform definition.

We only have 3 implementations right now:
- browser
- vscode
- android studio

Reviewed By: bolinfest

Differential Revision: D42589867

fbshipit-source-id: be2f3783b4d471fa9d503a9ff3d4464f7e5954ab