gituser: use ui.username() instead of direct config query
Summary:
I mainly want this so the HGUSER env var is respected in tests.

I prefer ui.username()'s abort message to what get_identity_or_raise() had since it cuts to the chase instead of "forcing" the user to open the link to discover the recommended command for setting username.

Reviewed By: bolinfest

Differential Revision: D42614255

fbshipit-source-id: de4a6719bfdefb2298e316bd18bad2c564aded84