indexedlog_store: fix for_loops_over_fallibles warning
Summary:
https://doc.rust-lang.org/nightly/nightly-rustc/rustc_lint/for_loops_over_fallibles/static.FOR_LOOPS_OVER_FALLIBLES.html

Both Option and Result implement IntoIterator trait, which allows using them in a for loop. for loop over Option or Result will iterate either 0 (if the value is None/Err(_)) or 1 time (if the value is Some(_)/Ok(_)). This is not very useful and is more clearly expressed via `if let`.

Reviewed By: muirdm

Differential Revision: D42656184

fbshipit-source-id: b14b53b49ef90eea26a2329dbe37168e5d454dea