changesets: Add unit test for add_many
Summary:
This adds tests to verify `Changesets::add_many` works as expected.

It also improves the changesets unit tests by using `paste` to make the macro simpler and adding `pretty_assertion`.

On the following diff I'll modify `add_many` and this test will make sure it continues working. I added it on this diff to prove it works with the current implementation.

Reviewed By: markbt

Differential Revision: D42635815

fbshipit-source-id: ada3aad775b8aa8565732244270b1a1a4badc18a