Move slice_repository to its own library
Summary:
The `slice_repository` function is used in derived data backfiller, to remove the need from loading the whole commit graph to start deriving. This is very useful and could be used in other places, so this diff extracts that behaviour and makes it more generic.

Done:
- Move to a different lib
- Deblobrepoify it
- Make it more generic, so it's not derived data specific, and can be used for anything you can say "still needs to be done on this commit", and depends on its parents being done.

I will use this library in another place on the next commit.

Reviewed By: markbt

Differential Revision: D42640268

fbshipit-source-id: 66bf297c6790dce6412c7f0a283bc0379fc125c3