Fetch application info from the server
Summary: In the next diff, we'll add a help dropdown for filing a bug. We want to show the ISL version number here. We don't have that info yet on the client, so we need to send it along. This requires the typical request+response messages like we do for other atoms.

Reviewed By: bolinfest

Differential Revision: D42732036

fbshipit-source-id: 73c03872583fb88ff15ef9f0aa14dfa65b43d6bd