Add <Copyable> component
Summary:
Add a component that handles clicking to copy the text content. It shows a copy icon on hover. After clicking to copy, it will show a success tooltip for a moment then auto-dismiss.

Such a component will be useful to have a consistent experience to click-to-copy text.

I think it's nice when you can just click to copy. Of course, I don't want to get in the way of good ole' selecting text + cmd-c. So that should still work here.

I added a small animation when you click to give some more tactile feedback.

I tried adding a hover tooltip to say "click to copy", but it kind of got in the way of the confirmation tooltip. Might be nice to still add that later.

Reviewed By: bolinfest

Differential Revision: D42732039

fbshipit-source-id: 2dbbb61bb72bde62878f8388c32720d86c1ccced