service: remove race between mount shutdown and Thrift calls
Summary:
The isSafeForInodeAccess is used to know when the root inode is initialized and
valid, but this is used in racy context. For instance, the getMountPoints
collects all the mounts whose root inode is initialized, but then releases the
lock on the mountPoints_ list, subsequently, the EdenMount's root Inode is
accessed without checking. If the EdenMount is shutdown prior to the root Inode
being copied, the root Inode might be a nullptr, causing a NULL-dereference.

This pattern can be found in all of the Thrift entry points that play with
mounts and thus are all potentially sources of EdenFS crashes. To solve this,
we need to guarantee 2 things:
 - isSafeForInodeAccess must only change state while the mountPoints_ lock is
   held, specifically at shutdown time,
 - The root inode must be copied while the mountPoints_ lock is held.

This diff does both. For the second one in particular EdenServer::getMount is
modified to also return the root Inode. This is done instead of introducing a new
function to prevent future bugs from seeping in that may be missed at review time.

Reviewed By: chadaustin

Differential Revision: D42563083

fbshipit-source-id: c0267277a54c425f330bbd58d1dc86ec3746502d