test showing issues with empty manifest
Summary:
If the commit without parents is also one of the roots of the repo it has null hg manifest. Many things in our infra are struggling with such commits, most recently monononoke_hg_sync job. This diff shows that the issue exists but doesn't event get to syncing commits because it fails to push those commits in the first place.

I'll use this test to show the impact of next few diffs in this stack.

Reviewed By: yancouto

Differential Revision: D42753350

fbshipit-source-id: 2995fb95b28343b85c8d08a3b7c14b2fcfa56c47