blobrepo: remove BlobRepo::get_blobstore
Summary:
Remove `BlobRepo::get_blobstore`, which returns a clone of the repo blobstore, and replace it with the equivalent facet methods.

In some cases, the obtained clone of the blobstore was just immediately followed by a call to `.boxed()`, which constructs a new `Arc` around the clone of the blobstore.  This is a double-indirection, and we can avoid this by using the `RepoBlobstoreArc` trait instead.

Differential Revision: D42747658

fbshipit-source-id: 4941068c7ccb3e4a627f4a26e0d2905e0b54d3a7