add sort_topological_starting_with_heads
Summary:
There are many valid topological orders for non-linear DAGs. When there are
merges the starting point for the traversal matters a lot. If we start from one
of the heads the merge branches would be continous parts of the sorted list.
If we start from the leaves we might end up interleaving merges.

This diff allows the topo sort caller to express preference about starting
point of the traversal. All nodes are still guaranteed to be traversed.

Reviewed By: markbt

Differential Revision: D42709711

fbshipit-source-id: ced83b813dc87a4157ed433cc0cddbed952819e0