remove needless addition of each node parents and dedup
Summary:
There's no need to add each node parents to the start points for DAG traversal
as they will be definitely visited when their parent is visited.

Now because we iterate only the keys list from dict there's no need to use
hashset for dedup. But.. even if there were duplicated on that list they'd be
harmless as they'd be skipped as "visited"

Reviewed By: markbt

Differential Revision: D42709712

fbshipit-source-id: 90ffdbf934e12d0b6be210be5ac2fb411c903adb