test-dirstate-nonnormalset: remove the test
Summary:
The test was intended to check that the nonnormalset is a subset of files
tracked by dirstate. Previously the nonnormalset is tracked separately so
it can be buggy and include files not in the main dirstate map. But the
TreeState format enforces the nonnormalset to be a subset of files tracked
by TreeState (dirstate). Therefore the test is no longer relevant.

Reviewed By: sggutier

Differential Revision: D42818816

fbshipit-source-id: 24277f4cd16efc82ac803e1ff2b5b53941289abd