status: enable rust status in test-fb-ext-tweakdefaults.t
Summary:
This test had a real failure where the new status wasn't respecting HGPLAIN to force root-relative paths in the output. It is safer to keep that behavior since we presume some automation would break if we changed to show cwd-relative paths unconditionally.

To use the native status I had to change this test to use modern repo format. That broke some use of revision numbers, so I tweaked tweakdefaults to respect the devel.default-date option to get stable commit hashes (so I could use those instead of revision numbers).

Reviewed By: quark-zju

Differential Revision: D42748900

fbshipit-source-id: 3358c659d5c67e781723720fcb2f760518d7d46f