Refactor Modal to consolidate styles
Summary:
Some changes to OptionsModal to make it easier to re-use in different cases and make the styling more consistent.

Context: <Modal> is a component you can render that is just the Modal rendering, without support for hiding / showing / dismissing / managing where it should be rendered from

<OptionsModal> is rendered as a result of the useOptionsModal API, which lets you trigger a custom content modal from anywhere more easily, and it manages visibility and closing for you.

Later in the stack, we'll extend/refactor <OptionsModal> to support more generic modals but with the nicer API that <OptionsModal> supports already.

 ---

Changes in this diff:

- make the dismiss 'x' part of `Modal` instead of `OptionsModal`, so it can be re-used easily
- add Icon as a prop to the OptionsModal, so we can be more consistent about rendering with icons like we do in dropdowns
- tweaks to styling for dismiss 'x', so it's absolutely positioned and doesn't move the title downwards
- tweaks to the only current use of OptionsModal, which is to set your submit command if you haven't used one before

Reviewed By: bolinfest

Differential Revision: D42764820

fbshipit-source-id: 067fb2efdb329430835d5131ebc4cdbb29f09c07