Refactor OptionsModal into UseModal
Summary:
We had a `useOptionsModal` utility, which creates a modal, adds it to a global state, then adds proper listeners for closing the modal on escape key, or if you click the 'x' in the top right, and sets up the right width for it, etc.

In this diff, I expand this to a more general `useModal` hook. This allows you to either use the more direct `{type: 'confirm', message: ..., buttons: ...}` API, or a more customizable `{type: 'custom', component: ...}` API. For file-a-bug, we'll use the more complicated API.

Without such a useModal util, right now you'd have to add a new <Modal> render on <App>, which needs its own global state management, then your own dismissing logic. This is sometimes needed, for example the Comparison View which does a lot of custom stuff.

But many smaller modals will fit a `useModal` pattern. For example, we may add a modal flow for Landing code, which would need custom rendering.

This API extension works like this:

```tsx
function SomeComponent() {
  const showModal = useModal();
  return <div>
    <button onClick={async () => {
       const result = await showModal({
         type: 'custom',
         component: MyModalRendererComponent,
         title: 'hello'
       });
       // now result is what the modal returned to us, which we can use for further processing
    }
  </div>
}

type MyData = {foo: string};

function MyModalRendererComponent({returnResultAndDismiss}: {returnResultAndDismiss: (data: MyData) => void}) {
  return <div>
    Custom modal content!
    <button onClick={() => {
      returnResultAndDismiss({foo: 'bar'});
    }}>Return some data</button>
  </div>;
}
```

We'll use this API later in the stack for file-a-bug, where the modal contains all the file-a-bug fields (title, description, ...),
and it returns that data to the code path where we open the modal.

Reviewed By: bolinfest

Differential Revision: D42818652

fbshipit-source-id: 4616690363f744762f6a46abf677123008127813