Dismiss bug tooltip when opening bug modal
Summary:
Later in this stack, I'll add a file-a-bug modal that shows up when you click the file-a-bug button in the dropdown. This modal should close the bug/help dropdown menu, or else it stays alive weirdly in the background.

To do this, we need to thread a `dismiss` prop through the various modal components.

Reviewed By: bolinfest

Differential Revision: D42764821

fbshipit-source-id: 6d561a53d04719e603ca6d2be0bf8ffee76bd731