add an uncached 4K random reads benchmark
Summary:
I had a benchmark in C++ to measure throughput of random reads, but
that was tricky to implement without a high-performance, portable file
abstraction.

So I wrote a similar benchmark in Rust. I wanted to do one pass across
the file, reading pages in random order, but I couldn't figure out how
to express that in Criterion. So, instead, read an arbitrary number of
pages, but disable caching with O_DIRECT on Linux, F_NOCACHE on macOS,
and FILE_FLAG_NO_BUFFERING on Windows.

Reviewed By: genevievehelsel

Differential Revision: D42650973

fbshipit-source-id: c315dfe348986d868b6360ea90529992199e5218