new trait for filtering the bundled changesets
Summary:
Right now we sync all the changesets that are not public commits on the server.
I'd like to add additional check that  will check if the changesests exist at
all in destination, in those cases it's fine to skip transferring them.

This diff is just building abstractions for adding that check. I'll add actual
checking logic in the next diff.

Differential Revision: D42851730

fbshipit-source-id: d8e98989340b4e4e70ad019fbf351e6ca1c2528d