sparse: ignore directory %includes
Summary:
Ingore %include directives that end in "/". It isn't valid to include a directory, and this was causing an error for the Rust status command when it went looking in the manifest for the path ending in "/".

We seem to sanitize the paths given to "hg sparse enable ...", so I'm not sure how the bad path got into a sparse config beyond hand editting.

Differential Revision: D42870518

fbshipit-source-id: d2162ca1180e06080c9bdf846f1715e389ea0fdd