test-debugrevset: fix debugruntest compat
Summary:
The test requires the Rust implementation of `debugrevset`. However by setting
`workingcopy.use-rust` to False (default) the `clidispatch` would request
falling back to a Python implementation of `debugrevset`, which does not exist.

Fix it by setting `workingcopy.use-rust` to True. We should make sure
`run-tests.py` and `debugruntest` has a same view of default hgrc. But that
is a potentially breaking change and would be done separately.

Reviewed By: zzl0

Differential Revision: D42825820

fbshipit-source-id: db537c0226d205f9fc5d1a464a6ef4b4d8793185