commitcloud: do not make LocalSerivce as a singleton
Summary:
If LocalSerivce is a singleton, it will capture an `ui` object that might be
stale in in-process hg tests. `LocalSerivce` is cheap to construct so let's
just avoid the singleton pattern for test compatibility.

This fixes `test-commitcloud-rename-workspace` debugruntest compatibility as
the following text is now printed properly using the up-to-date `ui.debugflag`:

  commitcloud local service: get_references for current version 2

Reviewed By: zzl0

Differential Revision: D42825822

fbshipit-source-id: 683dec345c4a36a6d79080e54a09050991c11dad