Support git_object_upload in Mononoke API
Summary: This method will later be used by the Git specific binary (i.e. SCS-like service for git purposes) to upload raw git objects (except file blobs) into Mononoke store. The git objects can be of type `Tree`, `Commit` or `Tag` but when stored in Mononoke the prefix will not capture the type of the object. The prefix for all objects would be `git_object_{object_id}`. The prefix will be preceeded by a repo-prefix since the blobstores (and multiplex) used exist at the repo-level. This is to ensure consistency between Mononoke and Git blobstore storage.

Differential Revision: D42927972

fbshipit-source-id: 4a003decd3144e9aed55eec31696100863e30a6f