Support debugging the isl server in vscode
Summary:
Add a vscode launch task so you can launch the ISL server `yarn serve` from inside vscode, and it will show the logs in the debug console, and work correctly with breakpoints and source maps.

This only works on the isl-server code. I tried adding a launch task for the client so it would spawn chrome with debugging enabled. This almost worked, except you had to paste the custom link in, and the source maps lines were inaccurate which made it pretty useless.

Also update our documentation with some more details about debugging and also how to use `yarn serve` with more arguments for clarity

Reviewed By: quark-zju

Differential Revision: D42906809

fbshipit-source-id: 1758fd954f863810a19e5e5539349d725a3c38aa