io: use a separate lock for quitting pager
Summary:
In the ctrlc handler we want to be able to quit the pager without being
blocked. However, that's not possible because the `Mutex<Inner>` struct that
has the logic to quit the pager can be already locked by a `write()` function
blocked by the pipe buffer.

This diff changes the `IO` state so it uses a separate lock to quit the pager.
If `write()` blocks, holding the `IOState` lock, it's still possible to obtain
the pager lock then send the `Quit` action to the pager.

Reviewed By: muirdm

Differential Revision: D42878476

fbshipit-source-id: 11f058e929d5e0ae72159b0a86a039ce4aa8f152