scs: fix topological sorting bug in "repo_stack_info" method
Summary:
The API says the `draft_commits` returned would be in topological order as per
https://fburl.com/code/a017qoam, but that does not seem to be the case.

Specifically, I created a stack of 4 commits as in:

```
╷
╷ o  f51d178c8  7 minutes ago  singhsrb
╷ │  Commit 4
╷ │
╷ o  73d322ffa  7 minutes ago  singhsrb
╷ │  Commit 3
╷ │
╷ o  74aca584a  7 minutes ago  singhsrb
╷ │  Commit 2
╷ │
╷ @  8b3703c4e  6 minutes ago  singhsrb
╭─╯  Commit 1
│
```

and called the API (https://fburl.com/thrift_fiddle/adrvdyf8) but the returned
draft commits were not in topological order. The `draft_commits` always lists
the passed in heads in order first which creates an issue if the heads are
itself not in topological order.

Reviewed By: mitrandir77

Differential Revision: D42968973

fbshipit-source-id: 992d9b4d2dde483595f62dd312494ef0af8a6376