cli: add --fast to doctor
Summary:
Some IDEs are running `eden doctor --dry-run` in the background to preemptively
find issues and display them to engineers. Unfortunately, some of the doctor
checks are fairly expensive and take a long time, these expensive checks are
also racy as they look at the working copy while it is being read/written,
which can cause more problems to be reported for non-existant issues.

To bypass both issues, a `--fast` option is introduced, which IDEs should use
to avoid the expensive checks.

Reviewed By: jdonald

Differential Revision: D43051918

fbshipit-source-id: 094c8db32ead99c9bfcd40cf1aad1e57c524b48a